% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuzzySolution.R
\name{fuzzySolution}
\alias{fuzzySolution}
\title{fuzzySolution}
\usage{
fuzzySolution(
  strata,
  cv,
  m = 2,
  minClusters = 2,
  maxclusters = 20,
  scale = TRUE,
  censiti = 0
)
}
\description{
fuzzySolution
}
