% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hEDA.R
\name{hEDA}
\alias{hEDA}
\title{hybrid Estimation of Distribution Algorithm (hEDA)}
\usage{
hEDA(
  stra,
  err,
  suggestions = NULL,
  Temp = 0.01,
  initialStr,
  decrement_constant = 0.95,
  end_time = 140,
  jsize = 10,
  length_of_markov_chain = 5,
  SAArun = TRUE,
  SAAiters = 1000,
  popSize = 200,
  iters = 100,
  mutationChance = NA,
  elitism = NA,
  addStrataFactor = 0.1,
  EDAfreq = EDAfreq,
  verbose = FALSE,
  dominio = dominio,
  minnumstrat = 2,
  kmax_percent = 0.025,
  ProbNewStratum = 1e-04,
  strcens = FALSE,
  writeFiles = FALSE,
  showPlot = TRUE,
  minTemp = 5e-06,
  realAllocation = TRUE
)
}
\description{
hybrid Estimation of Distribution Algorithm (hEDA)
}
