\name{HPK_SampleData}
\alias{HPK_SampleData}
\docType{data}
\title{
HPK_SampleData
}
\description{
Flow record of Cherry creek below Dion R Holm Powerhouse
}
\usage{data(HPK_SampleData)}
\format{

  \describe{
    \item{\code{site_no}}{identification number of gauge station}
    \item{\code{dataTime}}{Date and time flow when discharge recorded}
    \item{\code{dateTime}}{Date and time discharge recorded}
  }
}
\details{
A data frame with 3 columns and 30000 flow records
}
\source{
https://waterdata.usgs.gov/nwis/uv?site_no=11278400
}

\examples{
data(HPK_SampleData)
}

