% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KmeansSolution.R
\name{KmeansSolution}
\alias{KmeansSolution}
\title{KmeansSolution}
\usage{
KmeansSolution(
  strata,
  errors,
  nstrata = NA,
  minnumstrat = 2,
  maxclusters = NA,
  showPlot = TRUE
)
}
\description{
KmeansSolution
}
