% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KmeansSolution3.R
\name{KmeansSolution3}
\alias{KmeansSolution3}
\title{KmeansSolution3}
\usage{
KmeansSolution3(
  strata,
  errors,
  nstrata = NA,
  minnumstrat = 2,
  maxclusters = NA,
  showPlot = TRUE
)
}
\description{
KmeansSolution3
}
