% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KmedoidsSolution.R
\name{kmedoidsSolution}
\alias{kmedoidsSolution}
\title{kmedoidsSolution}
\usage{
kmedoidsSolution(
  strata,
  errors,
  nstrata = NA,
  minnumstrat = 2,
  maxclusters = NA,
  showPlot = TRUE
)
}
\description{
kmedoidsSolution
}
