% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallelHEDA.R
\name{parallelhEDA}
\alias{parallelhEDA}
\title{parallelhEDA}
\usage{
parallelhEDA(
  strata,
  cv,
  sugg,
  Temp = 1e-04,
  initialStrata = 15,
  decrement_constant = 0.95,
  end_time = 140,
  jsize = 5,
  length_of_markov_chain = 50,
  SAArun = TRUE,
  SAAiters = 50,
  popSize = 20,
  iters = 100,
  mutationChance = 0.01,
  elitism = 0.1,
  addStrataFactor = 1e-06,
  EDAfreq = 1,
  verbose = FALSE,
  dominio = dominio,
  minnumstrat = 2,
  kmax_percent = 0.025,
  ProbNewStratum = 1e-04,
  strcens = FALSE,
  writeFiles = FALSE,
  showPlot = TRUE,
  minTemp = 5e-06,
  realAllocation = TRUE
)
}
\description{
parallelhEDA
}
