% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SAA.R
\name{SAA}
\alias{SAA}
\title{Simulated Annealing Algorithm (SAA)}
\usage{
SAA(
  strata,
  errors,
  suggestions = NULL,
  Temp = NA,
  initialStrata,
  decrement_constant = 0.95,
  end_time = 140,
  showSettings = FALSE,
  jsize = 100,
  length_of_markov_chain = 50,
  verbose = FALSE,
  dominio = NULL,
  minnumstrat = NULL,
  kmax_percent = 0.025,
  ProbNewStratum = NA,
  strcens = FALSE,
  writeFiles = FALSE,
  showPlot = TRUE,
  minTemp = 5e-04,
  realAllocation = TRUE
)
}
\description{
Simulated Annealing Algorithm (SAA)
}
